+++ components/Flag/api/getFlagCountAPI.ts
@@ -2,14 +2,11 @@ import { Helpers } from "@quantfive/js-web-config";
 import API from "~/config/api";
 import { captureEvent } from "~/config/utils/events";
 
-export default function getFlagCountAPI(): Promise<void> {
-  return fetch(
-    API.FLAG_COUNT(),
-    API.GET_CONFIG()
-  )
+export default function getFlagCountAPI(): any {
+  return fetch(API.FLAG_COUNT(), API.GET_CONFIG())
     .then(Helpers.checkStatus)
     .then(Helpers.parseJSON)
-    .then((response:any) => {
+    .then((response: any) => {
       return response.count;
     })
     .catch((error: Error) => {
